cleanup done. inital app runs as normal. mapping not fulyl done. Ready to implement new data colelction app
