+++ streamlit/utils/helpers.py
@@ -1,0 +1,7 @@
+from pathlib import Path
+
+
+def read_markdown_file(markdown_file):
+    return Path(markdown_file).read_text()
+
+
